+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Update maildev instructions
 * Drop support for PHP < 8.1
 * Add some PHP tooling (PHP-CS-Fixer, PHPStan)
+* Add `castor init` command to initialize a new project
 
 ## 3.11.0 (2023-05-30)